+++ src/SendableError.ts
@@ -1,8 +1,8 @@
-import {v5 as uuid}                           from "uuid";
 import ErrorCode                              from "./ErrorCode";
 import {ErrorResponseBody, ResponseWithError} from "./Types";
 import {ERROR_CODE_MISC_INTERNAL_ERROR}       from "./DefaultCodes";
 import {getErrorLogger}                       from "./Logging";
+import crypto from "crypto";
 
 /*export const DEFAULT_ERROR_OPTIONS: Required<ErrorOptions> = {
   statusCode : 500,
@@ -10,12 +10,31 @@ import {getErrorLogger}                       from "./Logging";
   displayName: "Error",
 };*/
 
+const byteToHex: string[] = [];
+
+for (let i = 0; i < 256; ++i) {
+  byteToHex.push((i + 0x100).toString(16).substr(1));
+}
+
+function stringify(arr: any, offset = 0) {
+  // Note: Be careful editing this code!  It's been tuned for performance
+  // and works in ways you may not expect. See https://github.com/uuidjs/uuid/pull/434
+  const uuid = (byteToHex[arr[offset + 0]] + byteToHex[arr[offset + 1]] + byteToHex[arr[offset + 2]] + byteToHex[arr[offset + 3]] + '-' + byteToHex[arr[offset + 4]] + byteToHex[arr[offset + 5]] + '-' + byteToHex[arr[offset + 6]] + byteToHex[arr[offset + 7]] + '-' + byteToHex[arr[offset + 8]] + byteToHex[arr[offset + 9]] + '-' + byteToHex[arr[offset + 10]] + byteToHex[arr[offset + 11]] + byteToHex[arr[offset + 12]] + byteToHex[arr[offset + 13]] + byteToHex[arr[offset + 14]] + byteToHex[arr[offset + 15]]).toLowerCase(); // Consistency check for valid UUID.  If this throws, it's likely due to one
+  // of the following:
+  // - One or more input array values don't map to a hex octet (leading to
+  // "undefined" in the uuid)
+  // - Invalid input values for the RFC `version` or `variant` fields
+
+  return uuid;
+}
+
+
 /**
  * Transforms an error into a deterministic error identifier to enable easy log searching
  */
 export const getTraceId = (error: Error): string => {
   const errorString = error.stack || error.toString();
-  return uuid(errorString, uuid.URL);
+  return stringify(crypto.createHash("sha1").update(Buffer.from(errorString)).digest());
 };
 
 
@@ -41,6 +60,8 @@ export interface SendableErrorState {
   logged: boolean,
   traceId: string,
 }
+
+const CONSTRUCTOR_MESSAGE = "__$$sendable-error-message__";
 
 /**
  * An error with a known cause that is sendable
@@ -70,7 +91,7 @@ export default class SendableError<D extends SendableErrorDetails = {}> extends 
   }
 
   constructor(properties?: SendableErrorProperties<D>) {
-    super("__");
+    super(CONSTRUCTOR_MESSAGE);
 
     this.init(properties);
   }
@@ -97,6 +118,11 @@ export default class SendableError<D extends SendableErrorDetails = {}> extends 
     this.overrideProperty("code", this.getCode);
     this.overrideProperty("message", this.getMessage);
     this.overrideProperty("cause", this.getCause);
+    this.overrideProperty("stack", this.getStack);
+  }
+
+  getStack(): string | undefined {
+    return this.stack?.replace(CONSTRUCTOR_MESSAGE, this.getMessage());
   }
 
   getCause(): Error | undefined {